PICARD-2076: Use Windows long path prefix only, if long paths are unsupported
